Refactor | ProtectedRoute 컴포넌트 분리 및 페이지 라우팅 리팩토링
- ProtectedRoute 컴포넌트를 router 파일에서 분리하는 것이 관리하기 유용하다 생각되어 분리
- ProtectedRoute.tsx ) useEffect 파트 조건문 리팩토링
- router.tsx) 기본경로 index: true 로 설정
