+++ src/components/common/index.ts
@@ -1,3 +1,4 @@
+export { default as ProtectedRoute } from './ProtectedRoute';
 export { default as RootLayout } from './RootLayout';
 export { default as Button } from './Button';
 export { default as Checkbox } from './Checkbox';
