+++ src/routes/router.tsx
@@ -1,36 +1,7 @@
 import App from '@/App';
-import { useEffect } from 'react';
-import { createBrowserRouter, useNavigate, useParams } from 'react-router-dom';
+import { createBrowserRouter } from 'react-router-dom';
+import { ProtectedRoute } from '@/components/common';
 import { HomePage, SignInPage, SignUpPage, TodoPage } from '@/pages';
-
-interface ProtectedRouteProps {
-  element: JSX.Element;
-  path: string;
-}
-
-const isLoggedIn = () => {
-  const token = localStorage.getItem('token');
-  return !!token;
-};
-
-const ProtectedRoute = ({ element, path }: ProtectedRouteProps) => {
-  const navigate = useNavigate();
-  const params = useParams();
-
-  useEffect(() => {
-    if (path === 'signin' || path === 'signup') {
-      if (isLoggedIn()) {
-        navigate('/todo');
-      }
-    } else {
-      if (!isLoggedIn()) {
-        navigate('/signin');
-      }
-    }
-  }, [params]);
-
-  return element;
-};
 
 const router = createBrowserRouter([
   {
@@ -38,7 +9,7 @@ const router = createBrowserRouter([
     element: <App />,
     children: [
       {
-        path: '',
+        index: true,
         element: <ProtectedRoute element={<HomePage />} path="" />,
       },
       {
